pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "endstone-breeze"
version = "1.7.0"
description = "chat manager with extension support"
authors = [
    { name = "niko-at-chalupa"},
    { name = "Sasha"}
]
dependencies = [
    "alt-profanity-check",
    "wordfreq",
    "PyYAML"
]

[project.entry-points."endstone"]
breeze = "endstone_breeze:Breeze"

[tool.hatch.build.targets.wheel]
packages = ["src/endstone_breeze"]